refactor test support dependencies
fixes #25384
