+++ test/needle-api/needle-client-angular-generator.spec.ts
@@ -1,5 +1,5 @@
 import { before, it, describe } from 'esmocha';
-import { defaultHelpers as helpers, runResult } from '../support/index.js';
+import { defaultHelpers as helpers, runResult } from '../../testing/index.js';
 
 import AngularGenerator from '../../generators/angular/index.js';
 import { CLIENT_MAIN_SRC_DIR } from '../../generators/generator-constants.js';
